requirements.txt
spacy
PyPDF2
python-docx
pymongo
streamlit
streamlit
ollama
python-dotenv

python-dotenv


